backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30.0,<1.0
motor>=3.4.0,<4.0
pymongo>=4.7.0,<5.0
pydantic>=2.7.0,<3.0
email-validator>=2.1.0,<3.0
openai>=1.40.0,<2.0
python-dotenv>=1.0.1,<2.0
# Authentication dependencies
passlib[bcrypt]>=1.7.4,<2.0
PyJWT>=2.8.0,<3.0
bcrypt>=4.0.1,<5.0
